+++ assets/javascripts/kitten/karl/pages/partner/components/page.js
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react'
-import { Header } from '../../../../components/headers/header'
+import { HeaderNav } from '../../../../components/navigation/header-nav'
 import TextWithVideo from './text-with-video'
 import TextWithIconsList from './text-with-icon-list'
 import TextWithImages from './text-with-images'
@@ -7,10 +7,15 @@ import Title from './title'
 import TextWithImage from './text-with-image'
 import Triptych from './triptych'
 import FormWithImage from './form-with-image'
+import { KissKissBankBankLogo } from '../../../../components/logos/kisskissbankbanklogo'
 
 const PartnerPage = () => (
   <Fragment>
-    <Header />
+    <HeaderNav>
+      <HeaderNav.Logo href="#">
+        <KissKissBankBankLogo />
+      </HeaderNav.Logo>
+    </HeaderNav>
     <TextWithVideo />
     <TextWithIconsList />
     <TextWithImages />