+++ src/components/navigation/horizontal-nav/index.js
@@ -15,7 +15,7 @@ var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/h
 
 var _react = _interopRequireWildcard(require("react"));
 
-var _badge = require("../../../components/notifications/badge");
+var _badge = require("../../../components/atoms/badge");
 
 var _classnames = _interopRequireDefault(require("classnames"));