+++ src/esm/constants/colors-config.js
@@ -15,6 +15,8 @@ export default {
   // Semi Light Grey
   line2: '#d8d8d8',
   // Grey
+  line3: '#2d2d2d',
+  // Extra semi dark grey, not used for text
   primary1: '#19b4fa',
   // Primary
   primary2: '#05a8e6',
@@ -33,7 +35,7 @@ export default {
   // Semi Light Green
   valid: '#61d079',
   // Green
-  valid1: 'rgba(97, 208, 121, .1)',
+  valid1: 'hsl(133, 54%, 94%)',
   // Green Light
   error: '#ff0046',
   // Red
@@ -42,7 +44,7 @@ export default {
   error3: '#ffb2c7',
   // Semi Light Red
   orange: '#ff7800',
-  orange1: 'rgba(255, 130, 15, .1)',
+  orange1: '#fff2e7',
   // Orange light
   warning: '#8a6d3b',
   warning2: '#fcf8e3'